[SPARK-33614][SQL] Fix the constant folding rule to skip it if the expression fails to execute

+++ sql/catalyst/src/main/scala/org/apache/spark/sql/types/numerics.scala
@@ -145,7 +145,7 @@ private[sql] object FloatExactNumeric extends FloatIsFractional {
     if (Math.floor(x) <= longUpperBound && Math.ceil(x) >= longLowerBound) {
       x.toLong
     } else {
-      overflowException(x, "int")
+      overflowException(x, "long")
     }
   }
